fixed DOT numerical error
for reasons i cannot determine, the version with std::plus and std::multiply deduces as float when double is used,
which leads to a roundoff error.  using explicit typing fixes this.

the minimum solution is to use std::plus<double>() on the macro version but i added the explicit deduction
everywhere just to be safe.